[fix] comparison of constant '2' with boolean expression is always true
```c++
16872/include/seqan3/alphabet/detail/alphabet_base.hpp: In instantiation of 'constexpr derived_type& seqan3::alphabet_base<derived_type, size, char_t>::assign_rank(seqan3::alphabet_base<derived_type, size, char_t>::rank_type) [with derived_type = seqan3::mask; long unsigned int size = 2; char_t = void; seqan3::alphabet_base<derived_type, size, char_t>::rank_type = bool]':
/home/jenkins/ws/pipeline_header_tests/facb78a3/checkout-fd9b64940334c9f13f828ee16b0f4d95ec216872/include/seqan3/alphabet/mask/mask.hpp:92:51:   required from here
/home/jenkins/ws/pipeline_header_tests/facb78a3/checkout-fd9b64940334c9f13f828ee16b0f4d95ec216872/include/seqan3/alphabet/detail/alphabet_base.hpp:191:18: error: comparison of constant '2' with boolean expression is always true [-Werror=bool-compare]
         assert(c < value_size);
                ~~^~~
```

+++ include/seqan3/alphabet/detail/alphabet_base.hpp
@@ -188,7 +188,7 @@ public:
      */
     constexpr derived_type & assign_rank(rank_type const c) noexcept
     {
-        assert(c < value_size);
+        assert(static_cast<size_t>(c) < static_cast<size_t>(value_size));
         rank = c;
         return static_cast<derived_type &>(*this);
     }
